Add files via upload
Added init() function that calls setInterval() to changeCanvas() every 100ms (1/10th second). This eliminated the need for the two <form> elements hosting the inputs on the left and right side of pages. Originally, I used the two forms to handle changes to the canvas and redraw when an input changed. This was done with the css class ".reDrawOnChange" and a jquery function that called changeCanvas() when the class was changed, therefore updating the canvas. init() now handles this automatically every 100ms.