Add files via upload
Added undo last commit button and functionality. Undo will run savedCanvas.pop() to clear the last commit from the array, thus not having that item be drawn onto the main canvas anymore. It will undo just the last commit despite whichever screen (Text or image) that the user is on. So ie, saving text to the main canvas, then switching to the image tab and pressing undo will still clear the text.