Add files via upload
Bug fix where the temp canvas that holds text and image commits was double drawing over the main canvas, making text and text shadows appear unintentionally bolder and darker. Now this canvas is hidden when not in use, and visible when actively changing the background. Updated comments to reflect change.